Minor test code refactor (#24)
